@@ -1,7 +1,7 @@
 // see SignupForm.js for comments
 import React, { useState } from 'react';
 import { useMutation } from '@apollo/client';
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { Form, Button, Alert } from 'react-bootstrap';
 import '../assets/css/home.css';
 
@@ -10,10 +10,10 @@ import Auth from '../utils/auth';
 
 const LoginForm = () => {
   const [userFormData, setUserFormData] = useState({ email: '', password: '' });
-  const [login, {error}] = useMutation(LOGIN_USER);
+  const [login, { error }] = useMutation(LOGIN_USER);
   const [validated] = useState(false);
   const [showAlert, setShowAlert] = useState(false);
-  let history = useHistory();
+  let navigate = useNavigate();
 
   const handleInputChange = (event) => {
     const { name, value } = event.target;
@@ -22,26 +22,27 @@ const LoginForm = () => {
 
   const handleFormSubmit = async (event) => {
     event.preventDefault();
-
+    console.log(userFormData)
     // check if form has everything (as per react-bootstrap docs)
     const form = event.currentTarget;
     if (form.checkValidity() === false) {
       event.preventDefault();
       event.stopPropagation();
     }
-;
+    ;
     try {
-      const response = await login({
-        variables: {...userFormData},
+      const { data } = await login({
+        variables: { ...userFormData },
       });
-;
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
-      history.push("/entries");
-      const { token, user } = await response.json();
-      Auth.login(token);
-      
+      ;
+      // if (!response.ok) {
+      //   throw new Error('something went wrong!');
+      // }
+
+      // const { token, user } = await response.json();
+      Auth.login(data.login.token);
+      navigate("/entries");
+
     } catch (err) {
       console.error(error);
       setShowAlert(true);
