@@ -1,12 +1,7 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import {
-  Jumbotron,
   Container,
-  Col,
-  Form,
   Button,
-  Card,
-  CardColumns,
 } from "react-bootstrap";
 import { Link } from 'react-router-dom'
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -14,25 +9,21 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 const Home = () => {
   return (
     <Container className="d-flex p-2">
-      <Jumbotron>
-        <h1>C-est-la-vie</h1>
-        <p>
-          Welcome to C-est-la-vie, login or sign-up to add your journals and start your journey to self-care.
-        </p>
-        <p>
-          <Button type="submit" variant="success" size="lg">
-            <Link to="/login"> Login </Link>
-          </Button>
-          <Button type="submit" variant="success" size="lg">
-            <Link to="/signup"> Signup </Link>
-          </Button>
-        </p>
-      </Jumbotron>
-
+      <div className="container-fluid bg-light text-dark p-5">
+        <div className="container bg-light p-5">
+          <h1 className="display-4 fw-bold">C-est-la-vie</h1>
+          <p>Welcome to C-est-la-vie, login or sign-up to add your journals and start your journey to self-care.</p>
+          <p>
+            <Button className="m-3 p-3" type="submit" variant="primary" size="lg">
+              <Link className="text-light" to="/login"> Login </Link>
+            </Button>
+            <Button className="m-3 p-3" type="submit" variant="primary" size="lg">
+              <Link className="text-light" to="/signup"> Signup </Link>
+            </Button>
+          </p>
+        </div>
+      </div>
     </Container>
-
-
-
   )
 }
 
