@@ -3,61 +3,62 @@ const { AuthenticationError } = require("apollo-server-express");
 const { signToken } = require("../utils/auth");
 
 const resolvers = {
-    Query: {
-      me: async (parent, args, context) => {
-        if (context.user) {
-          const userData = await User.findOne({ _id: context.user._id }).select(
-            "-__v -password"
-          );
-          return userData;
-        }
-        throw new AuthenticationError("Not Logged In");
-      },
+  Query: {
+    me: async (parent, args, context) => {
+      if (context.user) {
+        const userData = await User.findOne({ _id: context.user._id }).select(
+          "-__v -password"
+        );
+        return userData;
+      }
+      throw new AuthenticationError("Not Logged In");
+    },
+  },
+
+  Mutation: {
+    addUser: async (parent, { username, email, password }) => {
+      const user = await User.create({ username, email, password });
+      const token = signToken(user);
+      return { token, user };
     },
 
-    Mutation: {
-        addUser: async (parent, { username, email, password }) => {
-          const user = await User.create({ username, email, password });
-          const token = signToken(user);
-          return { token, user };
-        },
-    
-        login: async (parent, { email, password }) => {
-          const user = await User.findOne({ email });
-          if (!user) {
-            throw new AuthenticationError("Incorrect credentials");
-          }
-          const correctPw = await user.isCorrectPassword(password);
-          if (!correctPw) {
-            throw new AuthenticationError("Incorrect credentials");
-          }
-          const token = signToken(user);
-          return { token, user };
-        },
-        saveEntries: async (parent, {input}, context) => {
-            if (context.user) {
-              const updatedUser = await User.findOneAndUpdate(
-                { _id: context.user._id },
-                { $push: { saveEntries: input } },
-                { new: true }
-              );
-              return updatedUser;
-            }
-            throw new Error("Could not save entry!");
-          },
-      
-          removeEntries: async (parent, { EntriesId }, context) => {
-            if (context.user) {
-              const updatedUser = await User.findOneAndUpdate(
-                { _id: context.user._id },
-                { $pull: { saveEntries: { EntriesId } } },
-                { new: true }
-              );
-              return updatedUser;
-            }
-            throw new AuthenticationError("Couldnt not delete entry!");
-          },
-        },
-      };
-      
-      module.exports = resolvers;
+    login: async (parent, { email, password }) => {
+      console.log(email)
+      const user = await User.findOne({ email });
+      if (!user) {
+        throw new AuthenticationError("Incorrect credentials");
+      }
+      const correctPw = await user.isCorrectPassword(password);
+      if (!correctPw) {
+        throw new AuthenticationError("Incorrect credentials");
+      }
+      const token = signToken(user);
+      return { token, user };
+    },
+    saveEntries: async (parent, { input }, context) => {
+      if (context.user) {
+        const updatedUser = await User.findOneAndUpdate(
+          { _id: context.user._id },
+          { $push: { saveEntries: input } },
+          { new: true }
+        );
+        return updatedUser;
+      }
+      throw new Error("Could not save entry!");
+    },
+
+    removeEntries: async (parent, { EntriesId }, context) => {
+      if (context.user) {
+        const updatedUser = await User.findOneAndUpdate(
+          { _id: context.user._id },
+          { $pull: { saveEntries: { EntriesId } } },
+          { new: true }
+        );
+        return updatedUser;
+      }
+      throw new AuthenticationError("Couldnt not delete entry!");
+    },
+  },
+};
+
+module.exports = resolvers;
